Finish uncompleted Route53 tasks on startup in background
