+++ dns/route53/operation.go
@@ -1,0 +1,9 @@
+package route53
+
+type operationType string
+
+const (
+	isDomainRegistered operationType = "IsDomainRegistered"
+	registerDomain     operationType = "RegisterDomain"
+	unregisterDomain   operationType = "UnregisterDomain"
+)
