+++ dns/route53/statestore.go
@@ -6,6 +6,7 @@ type awsRoute53StateStore interface {
 	create(state *domainState) error
 	update(state *domainState) error
 	find(orgId uint, domain string, state *domainState) (bool, error)
+	findByStatus(status string) ([]domainState, error)
 	listUnused() ([]domainState, error)
 	delete(state *domainState) error
 }
